requirements.txt
fastapi
hypercorn
python-multipart

faster-whisper
ffmpeg-python

nest-asyncio
pyngrok
